+++ src/Files.App/ViewModels/Settings/FoldersViewModel.cs
@@ -7,35 +7,16 @@ namespace Files.App.ViewModels.Settings
 	{
 		private IUserSettingsService UserSettingsService { get; } = Ioc.Default.GetRequiredService<IUserSettingsService>();
 
-		// FileTag combobox indexes (required to hide SyncStatus)
-		private readonly int FileTagSortingIndex = 5;
-		private readonly int FileTagGroupingIndex = 6;
+
 
 		public FoldersViewModel()
 		{
-			SelectedDefaultLayoutModeIndex = (int)DefaultLayoutMode;
-			SelectedDefaultSortingIndex = UserSettingsService.FoldersSettingsService.DefaultSortOption == SortOption.FileTag ? FileTagSortingIndex : (int)UserSettingsService.FoldersSettingsService.DefaultSortOption;
-			SelectedDefaultGroupingIndex = UserSettingsService.FoldersSettingsService.DefaultGroupOption == GroupOption.FileTag ? FileTagGroupingIndex : (int)UserSettingsService.FoldersSettingsService.DefaultGroupOption;
-			SelectedDefaultGroupByDateUnitIndex = (int)UserSettingsService.FoldersSettingsService.DefaultGroupByDateUnit;
-			SelectedDefaultSortPriorityIndex = UserSettingsService.FoldersSettingsService.DefaultSortDirectoriesAlongsideFiles ? 2 : UserSettingsService.FoldersSettingsService.DefaultSortFilesFirst ? 1 : 0;
 			SelectedDeleteConfirmationPolicyIndex = (int)DeleteConfirmationPolicy;
 		}
 
 		// Properties
 
-		private int selectedDefaultLayoutModeIndex;
-		public int SelectedDefaultLayoutModeIndex
-		{
-			get => selectedDefaultLayoutModeIndex;
-			set
-			{
-				if (SetProperty(ref selectedDefaultLayoutModeIndex, value))
-				{
-					OnPropertyChanged(nameof(SelectedDefaultLayoutModeIndex));
-					DefaultLayoutMode = (FolderLayoutModes)value;
-				}
-			}
-		}
+
 
 		private int selectedDeleteConfirmationPolicyIndex;
 		public int SelectedDeleteConfirmationPolicyIndex
@@ -47,105 +28,6 @@ namespace Files.App.ViewModels.Settings
 				{
 					OnPropertyChanged(nameof(SelectedDeleteConfirmationPolicyIndex));
 					DeleteConfirmationPolicy = (DeleteConfirmationPolicies)value;
-				}
-			}
-		}
-
-		public bool SyncFolderPreferencesAcrossDirectories
-		{
-			get => UserSettingsService.FoldersSettingsService.SyncFolderPreferencesAcrossDirectories;
-			set
-			{
-				if (value != UserSettingsService.FoldersSettingsService.SyncFolderPreferencesAcrossDirectories)
-				{
-					UserSettingsService.FoldersSettingsService.SyncFolderPreferencesAcrossDirectories = value;
-
-					ResetLayoutPreferences();
-					OnPropertyChanged();
-				}
-			}
-		}
-
-		public bool ShowFileTagColumn
-		{
-			get => UserSettingsService.FoldersSettingsService.ShowFileTagColumn;
-			set
-			{
-				if (value != UserSettingsService.FoldersSettingsService.ShowFileTagColumn)
-				{
-					UserSettingsService.FoldersSettingsService.ShowFileTagColumn = value;
-
-					OnPropertyChanged();
-				}
-			}
-		}
-
-		public bool ShowSizeColumn
-		{
-			get => UserSettingsService.FoldersSettingsService.ShowSizeColumn;
-			set
-			{
-				if (value != UserSettingsService.FoldersSettingsService.ShowSizeColumn)
-				{
-					UserSettingsService.FoldersSettingsService.ShowSizeColumn = value;
-
-					OnPropertyChanged();
-				}
-			}
-		}
-
-		public bool ShowTypeColumn
-		{
-			get => UserSettingsService.FoldersSettingsService.ShowTypeColumn;
-			set
-			{
-				if (value != UserSettingsService.FoldersSettingsService.ShowTypeColumn)
-				{
-					UserSettingsService.FoldersSettingsService.ShowTypeColumn = value;
-
-					OnPropertyChanged();
-				}
-			}
-		}
-
-		public bool ShowDateCreatedColumn
-		{
-			get => UserSettingsService.FoldersSettingsService.ShowDateCreatedColumn;
-			set
-			{
-				if (value != UserSettingsService.FoldersSettingsService.ShowDateCreatedColumn)
-				{
-					UserSettingsService.FoldersSettingsService.ShowDateCreatedColumn = value;
-
-					OnPropertyChanged();
-				}
-			}
-		}
-
-		public bool ShowDateColumn
-		{
-			get => UserSettingsService.FoldersSettingsService.ShowDateColumn;
-			set
-			{
-				if (value != UserSettingsService.FoldersSettingsService.ShowDateColumn)
-				{
-					UserSettingsService.FoldersSettingsService.ShowDateColumn = value;
-
-					OnPropertyChanged();
-				}
-			}
-		}
-
-		public FolderLayoutModes DefaultLayoutMode
-		{
-			get => UserSettingsService.FoldersSettingsService.DefaultLayoutMode;
-			set
-			{
-				if (value != UserSettingsService.FoldersSettingsService.DefaultLayoutMode)
-				{
-					UserSettingsService.FoldersSettingsService.DefaultLayoutMode = value;
-
-					OnPropertyChanged();
 				}
 			}
 		}
@@ -248,82 +130,6 @@ namespace Files.App.ViewModels.Settings
 			}
 		}
 
-		public bool SortInDescendingOrder
-		{
-			get => UserSettingsService.FoldersSettingsService.DefaultDirectorySortDirection == SortDirection.Descending;
-			set
-			{
-				if (value != (UserSettingsService.FoldersSettingsService.DefaultDirectorySortDirection == SortDirection.Descending))
-				{
-					UserSettingsService.FoldersSettingsService.DefaultDirectorySortDirection = value ? SortDirection.Descending : SortDirection.Ascending;
-					OnPropertyChanged();
-				}
-			}
-		}
-
-		public bool GroupInDescendingOrder
-		{
-			get => UserSettingsService.FoldersSettingsService.DefaultDirectoryGroupDirection == SortDirection.Descending;
-			set
-			{
-				if (value != (UserSettingsService.FoldersSettingsService.DefaultDirectoryGroupDirection == SortDirection.Descending))
-				{
-					UserSettingsService.FoldersSettingsService.DefaultDirectoryGroupDirection = value ? SortDirection.Descending : SortDirection.Ascending;
-					OnPropertyChanged();
-				}
-			}
-		}
-
-		public bool IsDefaultGrouped
-			=> UserSettingsService.FoldersSettingsService.DefaultGroupOption != GroupOption.None;
-
-		private int defaultGroupByDateUnitIndex;
-		public int SelectedDefaultGroupByDateUnitIndex
-		{
-			get => defaultGroupByDateUnitIndex;
-			set
-			{
-				if (SetProperty(ref defaultGroupByDateUnitIndex, value))
-				{
-					OnPropertyChanged(nameof(SelectedDefaultGroupByDateUnitIndex));
-					UserSettingsService.FoldersSettingsService.DefaultGroupByDateUnit = (GroupByDateUnit)value;
-				}
-			}
-		}
-
-		public bool IsGroupByDate
-			=> UserSettingsService.FoldersSettingsService.DefaultGroupOption.IsGroupByDate();
-
-		private int selectedDefaultSortPriorityIndex;
-		public int SelectedDefaultSortPriorityIndex
-		{
-			get => selectedDefaultSortPriorityIndex;
-			set
-			{
-				if (SetProperty(ref selectedDefaultSortPriorityIndex, value))
-				{
-					OnPropertyChanged(nameof(SelectedDefaultSortPriorityIndex));
-
-					switch (value)
-					{
-						case 0:
-							UserSettingsService.FoldersSettingsService.DefaultSortDirectoriesAlongsideFiles = false;
-							UserSettingsService.FoldersSettingsService.DefaultSortFilesFirst = false;
-							break;
-						case 1:
-							UserSettingsService.FoldersSettingsService.DefaultSortDirectoriesAlongsideFiles = false;
-							UserSettingsService.FoldersSettingsService.DefaultSortFilesFirst = true;
-							break;
-						case 2:
-							UserSettingsService.FoldersSettingsService.DefaultSortDirectoriesAlongsideFiles = true;
-							break;
-						default:
-							break;
-					}
-				}
-			}
-		}
-
 		public bool CalculateFolderSizes
 		{
 			get => UserSettingsService.FoldersSettingsService.CalculateFolderSizes;
@@ -348,40 +154,6 @@ namespace Files.App.ViewModels.Settings
 					UserSettingsService.FoldersSettingsService.ScrollToPreviousFolderWhenNavigatingUp = value;
 
 					OnPropertyChanged();
-				}
-			}
-		}
-
-		private int selectedDefaultSortingIndex;
-		public int SelectedDefaultSortingIndex
-		{
-			get => selectedDefaultSortingIndex;
-			set
-			{
-				if (SetProperty(ref selectedDefaultSortingIndex, value))
-				{
-					OnPropertyChanged(nameof(SelectedDefaultSortingIndex));
-
-					UserSettingsService.FoldersSettingsService.DefaultSortOption = value == FileTagSortingIndex ? SortOption.FileTag : (SortOption)value;
-				}
-			}
-		}
-
-		private int selectedDefaultGroupingIndex;
-		public int SelectedDefaultGroupingIndex
-		{
-			get => selectedDefaultGroupingIndex;
-			set
-			{
-				if (SetProperty(ref selectedDefaultGroupingIndex, value))
-				{
-					OnPropertyChanged(nameof(SelectedDefaultGroupingIndex));
-
-					UserSettingsService.FoldersSettingsService.DefaultGroupOption = value == FileTagGroupingIndex ? GroupOption.FileTag : (GroupOption)value;
-
-					// Raise an event for the grouping option toggle switches availability
-					OnPropertyChanged(nameof(IsDefaultGrouped));
-					OnPropertyChanged(nameof(IsGroupByDate));
 				}
 			}
 		}
@@ -482,14 +254,6 @@ namespace Files.App.ViewModels.Settings
 					OnPropertyChanged();
 				}
 			}
-		}
-
-		public void ResetLayoutPreferences()
-		{
-			// Is this proper practice?
-			var dbInstance = LayoutPreferencesManager.GetDatabaseManagerInstance();
-
-			dbInstance.ResetAll();
 		}
 	}
 }
