+++ src/Files.App/Views/Settings/LayoutPage.xaml.cs
@@ -1,0 +1,15 @@
+// Copyright (c) 2023 Files Community
+// Licensed under the MIT License. See the LICENSE.
+
+using Microsoft.UI.Xaml.Controls;
+
+namespace Files.App.Views.Settings
+{
+	public sealed partial class LayoutPage : Page
+	{
+		public LayoutPage()
+		{
+			InitializeComponent();
+		}
+	}
+}
